refactored profile image link into popup
