--- src/lib/index.ts
@@ -1,8 +1,0 @@
-// place files you want to import through the `$lib` alias in this folder.
-import Navbar from './components/Navbar.svelte';
-import PlayerBar from './components/PlayerBar.svelte';
-import Tracklist from './components/Tracklist.svelte';
-import Track from './components/Track.svelte';
-import LogoutButton from './components/LogoutButton.svelte';
-
-export { Navbar, PlayerBar, Tracklist, Track, LogoutButton };
