+++ src/routes/+layout.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
 	import { AppShell } from '@skeletonlabs/skeleton';
-	import { Navbar } from '$lib';
+	import AppBar from '$lib/components/AppBar.svelte';
 	import '../app.postcss';
+
 	// Popup imports
 	import { computePosition, autoUpdate, offset, shift, flip, arrow } from '@floating-ui/dom';
 	import { storePopup } from '@skeletonlabs/skeleton';
 	import type { LayoutData } from './$types';
-	import LogoutButton from '$lib/components/LogoutButton.svelte';
 	// Popup setup code
 	storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow });
 
@@ -15,15 +15,12 @@
 
 <AppShell>
 	<svelte:fragment slot="header">
-		<Navbar user={data.user} />
+		<AppBar user={data.user} />
 	</svelte:fragment>
 	<!-- (sidebarLeft) -->
 	<!-- (sidebarRight) -->
 	<!-- (pageHeader) -->
 	<!-- Router Slot -->
-	{#if data.user}
-		<LogoutButton />
-	{/if}
 	<slot />
 	<!-- ---- / ---- -->
 	<!-- (pageFooter) -->
