+++ src/routes/+page.svelte
@@ -1,9 +1,6 @@
 <script lang="ts">
-	import { PlayerBar } from '$lib';
 </script>
 
 <svelte:head>
 	<title>SpotifyVis</title>
 </svelte:head>
-
-<PlayerBar />
